Use greater of creationTimeStamp and lastTimestamp to determine event age
Currently, kubernetes-events-exporter looks at lastTimestamp (eventTime if lastTimestamp is missing)
of the event to determine the event age. If event age exceeded maxEventAgeSeconds, the event
is discarded. While based on event aggregation, lastTimestamp should be greater than creationTimestamp,
in practice it doesn't always seem to be the case. Hence, the current way of determining eventAge
sometimes results in events younger than maxEventAgeSeconds not getting processed. So as part of this
change, we use the greater of creationTimestamp and lastTimestamp (or eventTime if lastTimestamp is empty
and firstTimestamp if eventTime is empty) to calculate event age. This allows us to capture more
events rather than less and even if an event is processed more than once, we can de-dup it during
analysis by using a distinct operation on ResourceVersion.

**How event aggregation works at a high level**
At a high level, client-go maintains a local events cache. When a new event comes, it correlates
this event with the ones in cache. If it finds a matching event in the cache, it just updates the
event count and the message (prefixes message with "(combined from similar events) $original-message")
and patches the existing event instead of creating a new one. This is an optimization to reduce the
storage needed to persist events in ETCD. Additionally, events have a TTL set (defaults to 1hr). So,
if client-go tries to patch an event and the event has already expired, it creates a new event in this
case. In such scenarios, the lastTimestamp of the event is kept from what was in the cache and the
creationTimestamp is set by the apiserver when the event is created. While technically the delta between
lastTimestamp and creationTimestamp should be very small, in practice, the delta is seen up to a few minutes.
We haven't exactly found why that is the case but a probable reason is that before the event is created, the
request to apiserver is queued.

References:
* Here's an explanation of how event aggregation works - https://dev.to/shuheiktgw/kubernetes-event-aggregation-and-spam-filtering-in-client-go-25mn
* Event aggregation: https://github.com/kubernetes/kubernetes/blob/d9c7616d4b2f563c055df1f378d4dbeed33af869/staging/src/k8s.io/client-go/tools/record/events_cache.go#L242
* Event update: https://github.com/kubernetes/kubernetes/blob/d9c7616d4b2f563c055df1f378d4dbeed33af869/staging/src/k8s.io/client-go/tools/record/events_cache.go#L327